+++ drizzlepac/hapsequencer.py
@@ -54,9 +54,11 @@ import sys
 import traceback
 
 import numpy as np
+from astropy.io import fits
 from astropy.table import Table
 
 import drizzlepac
+from drizzlepac import util
 from drizzlepac.haputils import config_utils
 from drizzlepac.haputils import diagnostic_utils
 from drizzlepac.haputils import hla_flag_filter
@@ -67,8 +69,9 @@ from drizzlepac.haputils import svm_quality_analysis as svm_qa
 from drizzlepac.haputils.catalog_utils import HAPCatalogs
 
 from stsci.tools import logutil
+from stwcs import updatewcs
 from stwcs import wcsutil
-
+from stwcs.wcsutil import headerlet
 
 __taskname__ = 'hapsequencer'
 MSG_DATEFMT = '%Y%j%H%M%S'
@@ -92,32 +95,32 @@ envvar_cat_str = "SVM_CATALOG_{}"
 
 # --------------------------------------------------------------------------------------------------------------
 
+
 def create_catalog_products(total_obj_list, log_level, diagnostic_mode=False, phot_mode='both',
                             catalog_switches=None):
     """This subroutine utilizes haputils/catalog_utils module to produce photometric sourcelists for the specified
     total drizzle product and it's associated child filter products.
 
-    Parameters
-    ----------
-    total_obj_list : drizzlepac.haputils.Product.TotalProduct
-        total drizzle product that will be processed by catalog_utils. catalog_utils will also create photometric
-        sourcelists for the child filter products of this total product.
-
+    PARAMETERS
+    -----------
+    total_obj_list : `drizzlepac.haputils.product.TotalProduct`
+                    total drizzle product that will be processed by catalog_utils.
+                    catalog_utils will also create photometric
+                    sourcelists for the child filter products of this total product.
     log_level : int, optional
-        The desired level of verboseness in the log statements displayed on the screen and written to the .log file.
-
+                The desired level of verboseness in the log statements displayed on the screen and written to the .log file.
     diagnostic_mode : bool, optional
-        generate ds9 region file counterparts to the photometric sourcelists? Default value is False.
-
+                      generate ds9 region file counterparts to the photometric sourcelists? Default value is False.
     phot_mode : str, optional
-        Which algorithm should be used to generate the sourcelists? 'aperture' for aperture photometry;
-        'segment' for segment map photometry; 'both' for both 'segment' and 'aperture'. Default value is 'both'.
-
+                Which algorithm should be used to generate the sourcelists? 'aperture' for aperture photometry;
+                'segment' for segment map photometry; 'both' for both 'segment' and 'aperture'. Default value is 'both'.
     catalog_switches : dict, optional
-        Specify which, if any, catalogs should be generated at all, based on detector.  This dictionary
-        needs to contain values for all instruments; namely:
-        SVM_CATALOG_HRC, SVM_CATALOG_SBC, SVM_CATALOG_WFC, SVM_CATALOG_UVIS, SVM_CATALOG_IR
-        These variables can be defined with values of 'on'/'off'/'yes'/'no'/'true'/'false'.
+                       Specify which, if any, catalogs should be generated at all, based on detector.  This dictionary
+                       needs to contain values for all instruments; namely:
+
+                       SVM_CATALOG_HRC, SVM_CATALOG_SBC, SVM_CATALOG_WFC, SVM_CATALOG_UVIS, SVM_CATALOG_IR
+
+                       These variables can be defined with values of 'on'/'off'/'yes'/'no'/'true'/'false'.
 
     Returns
     -------
@@ -305,6 +308,7 @@ def create_catalog_products(total_obj_list, log_level, diagnostic_mode=False, ph
                     filter_product_catalogs.catalogs[cat_type].subset_filter_source_cat
 
             # ...and append the filter columns to the total detection product catalog.
+            log.info("create catalogs. combine() columns: {}".format(subset_columns_dict))
             total_product_catalogs.combine(subset_columns_dict)
 
         # At this point the total product catalog contains all columns contributed
@@ -371,14 +375,14 @@ def create_drizzle_products(total_obj_list):
 
     Parameters
     ----------
-    total_obj_list: list
+    total_obj_list : list
         List of TotalProduct objects, one object per instrument/detector combination is
         a visit.  The TotalProduct objects are comprised of FilterProduct and ExposureProduct
         objects.
 
     RETURNS
     -------
-    product_list: list
+    product_list : list
         A list of output products
     """
     log.info("Processing with astrodrizzle version {}".format(drizzlepac.astrodrizzle.__version__))
@@ -468,6 +472,7 @@ def create_drizzle_products(total_obj_list):
 
 # ----------------------------------------------------------------------------------------------------------------------
 
+
 def run_hap_processing(input_filename, diagnostic_mode=False, input_custom_pars_file=None,
                        output_custom_pars_file=None, phot_mode="both", log_level=logutil.logging.INFO):
     """
@@ -528,14 +533,14 @@ def run_hap_processing(input_filename, diagnostic_mode=False, input_custom_pars_
     cat_switches = {sw: _get_envvar_switch(sw, default=envvar_cat_svm[sw]) for sw in envvar_cat_svm}
 
     total_obj_list = []
-    product_list = []
     try:
         # Parse the poller file and generate the the obs_info_dict, as well as the total detection
         # product lists which contain the ExposureProduct, FilterProduct, and TotalProduct objects
         # A poller file contains visit data for a single instrument.  The TotalProduct discriminant
         # is the detector.  A TotalProduct object is comprised of FilterProducts and ExposureProducts
         # where its FilterProduct is distinguished by the filter in use, and the ExposureProduct
-        # is the atomic exposure data.
+        # is the atomic exposure data. Note: the TotalProduct was enhanced to also be comprised
+        # of an GrismExposureProduct list which is exclusive to the TotalProduct.
         log.info("Parse the poller and determine what exposures need to be combined into separate products.\n")
         obs_info_dict, total_obj_list = poller_utils.interpret_obset_input(input_filename, log_level)
 
@@ -553,6 +558,7 @@ def run_hap_processing(input_filename, diagnostic_mode=False, input_custom_pars_
         # Update all of the product objects with their associated configuration information.
         for total_item in total_obj_list:
             log.info("Preparing configuration parameter values for total product {}".format(total_item.drizzle_filename))
+
             total_item.configobj_pars = config_utils.HapConfig(total_item,
                                                                log_level=log_level,
                                                                input_custom_pars_file=input_custom_pars_file,
@@ -573,9 +579,24 @@ def run_hap_processing(input_filename, diagnostic_mode=False, input_custom_pars_
 
             log.info("The configuration parameters have been read and applied to the drizzle objects.")
 
-            reference_catalog = run_align_to_gaia(total_item, log_level=log_level, diagnostic_mode=diagnostic_mode)
-            if reference_catalog:
-                product_list += reference_catalog
+            # Check to ensure the object contains a list of direct images (protects against processing Grism/Prism data)
+            if total_item.edp_list:
+                reference_catalog = run_align_to_gaia(total_item, log_level=log_level, diagnostic_mode=diagnostic_mode)
+                if reference_catalog:
+                    product_list += reference_catalog
+
+            # Need to delete the Ramp filter Exposure objects from the *Product lists as
+            # these images will not be processed beyond the alignment to Gaia (run_align_to_gaia).
+            delete_ramp_exposures(total_item.fdp_list)
+            delete_ramp_exposures(total_item.edp_list)
+  
+        # If there are Grism/Prism images present in this visit, as well as corresponding direct images
+        # for the same detector, update the primary WCS in the direct and/or Grism/Prism images as
+        # appropriate to be an 'a priori' or the pipeline default (fallback) solution
+        for total_item in total_obj_list:
+            if total_item.grism_edp_list and total_item.edp_list:
+                grism_flt_list = update_wcs_in_visit(total_item)
+                product_list += grism_flt_list
 
         # Run AstroDrizzle to produce drizzle-combined products
         log.info("\n{}: Create drizzled imagery products.".format(str(datetime.datetime.now())))
@@ -637,6 +658,9 @@ def run_hap_processing(input_filename, diagnostic_mode=False, input_custom_pars_
         if total_obj_list:
             for tot_obj in total_obj_list:
                 proc_utils.append_trl_file(tot_obj.trl_filename, logname, clean=False)
+                # Update DRIZPARS keyword value with new logfile name in ALL drizzle products
+                tot_obj.update_drizpars()
+
         # Now remove single temp log file
         if os.path.exists(logname):
             os.remove(logname)
@@ -681,17 +705,22 @@ def run_align_to_gaia(tot_obj, log_level=logutil.logging.INFO, diagnostic_mode=F
     log.info("\n{}: Finished aligning gaia_obj to GAIA".format(str(datetime.datetime.now())))
     log.info("ALIGNED WCS: \n{}".format(tot_obj.meta_wcs))
 
-    # Return the name of the alignment catalog
+    # Clean up if the align_table does not exist - the metawcs file
+    # may already be deleted, but make sure here.
     if align_table is None:
         gaia_obj.refname = None
         headerlet_filenames = []
+        try:
+            os.remove(gaia_obj.refname)
+        except OSError:
+            pass
     else:
         # Get names of all headerlet files written out to file
         headerlet_filenames = [f for f in align_table.filtered_table['headerletFile'] if f != "None"]
 
-    return [gaia_obj.refname] + headerlet_filenames
-
+    return headerlet_filenames
 # ----------------------------------------------------------------------------------------------------------------------
+
 
 def run_sourcelist_flagging(filter_product_obj, filter_product_catalogs, log_level, diagnostic_mode=False):
     """
@@ -815,3 +844,331 @@ def _get_envvar_switch(envvar_name, default=None):
         switch_val = envvar_bool_dict[default] if default else None
 
     return switch_val
+
+# ------------------------------------------------------------------------------
+
+
+def update_wcs_in_visit(tdp):
+    """Examine entries in the total data product for Grism/Prism data
+
+    This routine is only invoked if the total data product for a particular detector
+    for the visit is comprised of both an ExposureProduct list AND a
+    GrismExposureProduct list which contain entries.
+
+    *** If the visit contains Grism/Prism data, then either the 'a priori' WCS for
+    the Grism/Prism data or the pipeline default WCS for the direct images will be
+    set as the primary/active WCS for all the Grism/Prism and direct image data for
+    that detector - whichever WCS is common to all the images.
+
+    Note: The direct exposure image active WCS solutions will be modified in-place.
+
+    Parameters
+    ----------
+    tdp : total data product (TotalProduct) object
+        Object comprised of FilterProduct, ExposureProduct, and GrismExposureProduct
+
+    Returns
+    -------
+    grism_product_list : list
+        List of all the SVM Grism/Prism FLT/FLC files updated with the common WCS
+
+    """
+    log.info("\n***** Grism/Prism Image Processing *****")
+    # The TotalProduct (tdp) for this instrument/dectector has both a Grism/Prism
+    # exposure list and a direct exposure list - both with contents.
+    # Every image should (!) have an IDC_?????????-GSC240 solution.
+    wcs_preference = ['IDC_?????????-GSC240', 'IDC_?????????', 'OPUS']
+
+    # Grism output product list for the manifest
+    grism_product_list = []
+
+    grism_wcs_set, skip_grism_list, g_keyword_wcs_names_dict, grism_dict = collect_wcs_names(tdp.grism_edp_list, 'GRISM')
+    log.info("WCS solutions common to all viable Grism/Prism images: {}".format(grism_wcs_set))
+
+    # There is a preference for the active WCS for the viable images in the visit
+    # Check the grism_wcs_set for the preferential solutions
+    # If a common name is found here, wait to report it as the active WCS may still need to
+    # be rolled back depending upon the WCS solutions available in the direct images.
+    match_list = []
+    grism_wcsname = ''
+    for wcs_item in wcs_preference:
+        match_list = fnmatch.filter(grism_wcs_set, wcs_item)
+        if match_list:
+            grism_wcsname = match_list[0]
+            log.info("Proposed WCS for use after Grism/Prism examination: {}".format(grism_wcsname))
+            break
+
+    if not grism_wcsname:
+        log.error("None of the preferred WCS names are present in the common set of WCS names for the Grism/Prism images.")
+        log.error("    There is a problem with this visit.  Deleting all SVM Grism/Prism FLT/FLC files.")
+        try:
+            for image_file in tdp.grism_edp_list.full_filename:
+                os.remove(image_file)
+                log.warning("Deleted Grism/Prism image {}.".format(image_file))
+        except OSError:
+            pass
+        sys.exit(1)
+
+        log.info("\nProposed WCS for use after Grism/Prism examination: {}".format(grism_wcsname))
+
+    # If it is now known there is at least one viable Grism/Prism observation, it is desired to make
+    # the active Grism/Prism WCS be the same for all of the Grism/Prism as well as the direct images
+    # in the visit.  The catch can be if any of the direct images do not have an 'a priori' solution
+    # which matches the Grism/Prism images.  In this case, it will be necessary to "fall back" to
+    # the pipeline WCS solution.
+    log.info("\n***** Direct Image Processing *****")
+
+    # Loop over all the direct images for this detector in the visit to update the WCS
+    for edp in tdp.edp_list:
+        filename = edp.full_filename
+        hdu = fits.open(filename)
+
+        # Make sure the WCS is up-to-date - doing the update for the
+        # direct images here so there is no need to modify a
+        # pre-existing class which could cause an issue
+        drizcorr = hdu[0].header['DRIZCORR']
+        if drizcorr == "OMIT":
+            updatewcs.updatewcs(filename, use_db=True)
+        hdu.close()
+
+    direct_wcs_set, skip_direct_list, d_keyword_wcs_names_dict, direct_dict = collect_wcs_names(tdp.edp_list, 'DIRECT')
+    log.info("WCS solutions common to all viable direct images: {}".format(direct_wcs_set))
+
+    # Are the grism_wcs_set and the direct_wcs_set disjoint?  If they are disjoint, there can
+    # be no common WCS solution.
+    is_disjoint = grism_wcs_set.isdisjoint(direct_wcs_set)
+    if not is_disjoint:
+        # Generate the intersection between the Grism/Prism and direct images
+        grism_wcs_set &= direct_wcs_set
+        log.info("Common WCS solutions exist between the Grism/Prism and the direct images.")
+        log.info("    The common WCS solutions are: {}".format(grism_wcs_set))
+
+        # There is a preference for the active WCS for the viable images in the visit
+        # Check the grism_wcs_set for the preferential solutions
+        match_list = []
+        final_wcsname = ''
+        for wcs_item in wcs_preference:
+            match_list = fnmatch.filter(grism_wcs_set, wcs_item)
+            if match_list:
+                final_wcsname = match_list[0]
+                log.info("Final WCS solution to use for all Grism/Prism and direct images: {}".format(final_wcsname))
+                break
+
+        if final_wcsname:
+            # Finally, if the image is not in a skip list, reset the primary WCS in all the images
+
+            for g_edp in tdp.grism_edp_list:
+                filename = g_edp.full_filename
+                if filename not in skip_grism_list:
+                    log.info("Setting the primary WCS for Grism/Prism image {} to {}.".format(filename, final_wcsname))
+                    update_active_wcs(filename, final_wcsname, g_keyword_wcs_names_dict, grism_dict)
+
+                    # Add the Grism/Prism images to the manifest as all of the files exist.
+                    grism_product_list.append(filename)
+
+            for edp in tdp.edp_list:
+                filename = edp.full_filename
+                if filename not in skip_direct_list:
+                    log.info("Setting the primary WCS for direct image {} to {}.".format(filename, final_wcsname))
+                    update_active_wcs(filename, final_wcsname, d_keyword_wcs_names_dict, direct_dict)
+        else:
+            # Do nothing
+            pass
+
+    # Disjoint
+    else:
+        log.info("The Grism/Prism and direct images in this visit have no WCS solution common to both sets of data.")
+        log.info("The active WCS solution for each Grism/Prism and direct image is not changed.")
+
+    log.info("Grism product list: {}".format(grism_product_list))
+    return grism_product_list
+
+# ------------------------------------------------------------------------------
+
+
+def collect_wcs_names(edp_list, image_type):
+    """
+    Utility to collect all the WCS solution names common to the input image list
+
+    Parameters
+    ----------
+    edp_list: str list
+        List containing the SVM FLT/FLC filenames
+
+    image_type: string
+        String containing either 'GRISM' or 'DIRECT' to use as a switch for
+        output information
+
+    Returns
+    -------
+    image_wcs_set: set of WCS solutions
+        The set contains the WCS solution names common to all of the
+        input exposures
+
+    skip_image_list: list
+        This is a list of exposures in the input list which should be
+        skipped/ignored when updating the active WCS solution
+
+    keyword_wcs_names_dict: dictionary {filename: list}
+        The dictionary is used to associate an individual image/filename with
+        a list of WCS solution names in the file stored as keywords (not headerlets)
+
+    image_dict: dictionary {filename: list}
+        The dictionary is used to associate an individual image/filename with
+        a list of *all* WCS solution names in the file 
+    
+    """
+
+    image_wcs_set = set()
+    skip_image_list = []
+    exist_image_set = False
+    image_dict = {}
+    keyword_wcs_names_dict = {}
+    # Loop over all the Grism/Prism images for this detector in the visit
+    for edp in edp_list:
+
+        filename = edp.full_filename
+
+        # Get all the WCS names which are common to all of the images
+        # Note that WCS solutions may be represented as FITS keyword values in the
+        # SCI extension and/or as headerlets in the HDRLET extensions.  
+        # Get the keyword WCS solution names.
+        keyword_wcs_names = list(wcsutil.altwcs.wcsnames(filename, ext=1).values())
+
+        # Get the headerlet WCS solution names
+        headerlet_wcs_names = wcsutil.headerlet.get_headerlet_kw_names(filename, kw="WCSNAME")
+        all_wcs_names = keyword_wcs_names + headerlet_wcs_names
+        keyword_wcs_names_dict[filename] = keyword_wcs_names
+        image_dict[filename] = all_wcs_names
+        if all_wcs_names:
+            log.debug("WCS solutions for file {} are {}.".format(filename, all_wcs_names))
+            # Initialize a set with wcsnames
+            if not exist_image_set:
+                image_wcs_set = set(all_wcs_names)
+                exist_image_set = True
+            # Generate the intersection with the set of existing wcsnames and the wcsnames
+            # from the current image
+            else:
+                image_wcs_set &= set(all_wcs_names)
+
+            # Oops...no common wcsnames
+            if not image_wcs_set:
+                log.error("There are no common WCS solutions with this image {} and previously processed images".format(filename))
+                log.error("    There is a problem with this image/visit.")
+                log.error("    Make sure the input data are not *_raw.fits files.")
+                sys.exit(1)
+        # If there are no WCS solutions, the image could be bad (e.g., EXPTIME=0 or EXPFLAG="TDF-DOWN...")
+        else:
+            log.warning("There are no WCS solutions in the image {} in this visit.".format(filename))
+            skip_image_list.append(filename)
+            if image_type == 'GRISM':
+                log.warning("    Skip and delete this image.")
+                # Delete the SVM FLT/FlC Grism/Prism image as it has no updated WCS 
+                try:
+                    os.remove(filename)
+                    log.warning("Deleted Grism/Prism image {}.".format(filename))
+                except OSError:
+                    pass
+            else:
+                log.warning("    Skip this image.")
+
+    return image_wcs_set, skip_image_list, keyword_wcs_names_dict, image_dict
+
+# ------------------------------------------------------------------------------
+
+
+def update_active_wcs(filename, wcsname, keyword_wcs_names_dict, image_dict):
+    """
+    Utility to update the active/primary WCS solution
+
+    This small utility updates the active/primary WCS solution for the input
+    file with the WCS solution indicted by the input parameter "wcsname"
+
+    Parameters
+    ----------
+    filename : str
+        Input/Output SVM FLT/FLC filename - the file is updated in-place
+
+    wcsname : str
+        Name of the desired WCS active/primary solution to be set for the filename
+
+    keyword_wcs_names_dict : dictionary associated with each image in visit
+        The dictionary is {filename: List of WCS solution names stored as keywords in SCI}
+
+    image_dict : dictionary associated with each image in visit
+        The dictionary is {filename: List of WCS solution names}
+
+    Returns
+    -------
+    None
+    
+    """
+    # For exposures with multiple science extensions (multiple chips),
+    # generate a combined WCS
+    num_sci_ext, extname = util.count_sci_extensions(filename)
+    extname_list = [(extname, x+1) for x in range(num_sci_ext)]
+
+    hdu = fits.open(filename, mode="update")
+
+    # Check if the desired WCS solution is already the active solution
+    # whereupon there is nothing to do
+    key = wcsutil.altwcs.getKeyFromName(hdu[1].header, wcsname)
+    if key != " ":
+
+        # If the current active solution is not already saved in the file, it
+        # must be stored as an alternate solution
+        wcs_list = image_dict[filename]
+        is_stored = [item for item in wcs_list if wcsname == item]
+        if not is_stored:
+            out_tuple = wcsutil.altwcs.archive_wcs(filename, extname_list, wcsname = wcsname)
+            log.info("Archiving previous active WCS solution: {}".format(out_tuple))
+
+        # Is the source of the wcsname for this image from keywords or a headerlet?
+        # The source dictates how the WCS will be made the active WCS
+        # Examine the "keyword" WCS solutions as a full string match...
+        keyword_wcs_list = keyword_wcs_names_dict[filename]
+        found_string = [i for i in keyword_wcs_list if wcsname == i]
+        if found_string:
+            wcsutil.altwcs.restoreWCS(filename, ext=extname_list, wcsname=found_string[0])
+        #...the headerlet WCS solutions -- need to get the HDRNAME to retrieve the headerlet
+        else:
+            headerlet_hdr_names = wcsutil.headerlet.get_headerlet_kw_names(filename, kw="HDRNAME")
+            found_string = [i for i in headerlet_hdr_names if wcsname in i]
+            wcsutil.headerlet.restore_from_headerlet(filename, hdrname=found_string[0], force=True)
+    else:
+        log.info("No need to update active WCS solution of {} for {} as it is already the active solution.".format(wcsname, filename))
+
+    hdu.close()
+
+# ------------------------------------------------------------------------------
+
+
+def delete_ramp_exposures(obj_list):
+    """Delete the Ramp filter objects from the Total Product internal lists
+
+    The Total Data Product (tdp) object is comprised of a list of Filter Product objects,
+    as well as a list of Exposure Product objects.  The Ramp filter
+    images need to be processed in the same manner as nominal exposures for at least
+    some of the processing steps.  Because of this, it was deemed best to keep the 
+    Ramp exposures in the tdp list attributes, until their final processing
+    stage (align_to_gaia), and then delete these objects from the attribute
+    lists. This function handles the deletion of Ramp images from the input
+    list.
+
+    Parameters
+    ----------
+    obj_list : list of either FilterProduct ro ExposureProduct objects
+        The list is updated in-place.
+
+    Returns
+    -------
+    None
+
+    """
+    temp = []
+    while obj_list:
+        obj = obj_list.pop()
+        if not obj.filters.lower().startswith('fr'):
+            temp.append(obj)
+    while temp:
+        obj_list.append(temp.pop())