+++ CHANGELOG.rst
@@ -18,8 +18,10 @@ of the list).
 ==================
 
 - Fix a bug in ``tweakback`` that may cause incorrect "updated" WCS to be
-  picked up from the drizzled image. [#913]3.1.9 (unreleased)
+  picked up from the drizzled image. [#913]
 
+- Added ``DRIZPARS`` keyword to final output drizzle product primary header
+  to document the name of the associated trailer file. [#934]
 
 3.2.0 (7-Dec-2020)
 ==================
@@ -38,7 +40,7 @@ mosaic processing used to create the single-visit mosaics products.
 
 
 3.1.8 (11-Aug-2020)
-==================
+===================
 
 A number of changes have been implemented to either correct problems or
 improve the processed results.  The most significant of the changes are: